backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
greenlet==3.1.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
cryptography==44.0.0

# HTTP client for 3CX
httpx==0.28.1
beautifulsoup4==4.12.3

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Utilities
python-dateutil==2.9.0
croniter==5.0.1
